Move metrics::tls_config_reload into telemetry
Following on #52, this change moves the
`telemetry::metrics::tls_config_reload` module to
`telemetry::tls_config_reload`.

The `Fmt` type has been renamed to `Report`.

Furthermore, two helper methods have been added to `metrics::Scopes` so
that its internal representation need not be made public.
